move validation os images to storage conftest
